requirement.txt.txt
tensorflow==2.12.0
keras==2.12.0
numpy==1.23.5
pandas==1.5.3
matplotlib==3.7.1
seaborn==0.12.2
nltk==3.8.1
scikit-learn==1.2.2
opencv-python==4.7.0.72
Pillow==9.4.0
tqdm==4.65.0
h5py==3.8.0
streamlit==1.24.0  # If deploying as a web app
